requirements.txt
numpy
flask
flask_cors
woocommerce
datetime
PyMySQL